LPS-112205 Move to different package to avoid packages split error
